+++ Packages/com.texelsaur.video/Editor/Inspectors/AudioManagerInspector.cs
@@ -27,10 +27,12 @@ namespace Texel
         SerializedProperty masterVolumeProperty;
         SerializedProperty masterMuteProperty;
         //SerializedProperty master2DProperty;
-        SerializedProperty masterVolumePersistenceProperty;
 
         SerializedProperty channelGroupsProperty;
         SerializedProperty defaultChannelGroupProperty;
+
+        SerializedProperty persistenceKeyProperty;
+        SerializedProperty masterVolumePersistenceProperty;
 
         SerializedProperty debugLogProperty;
         SerializedProperty debugStateProperty;
@@ -81,10 +83,12 @@ namespace Texel
             masterVolumeProperty = serializedObject.FindProperty(nameof(AudioManager.masterVolume));
             masterMuteProperty = serializedObject.FindProperty(nameof(AudioManager.masterMute));
             //master2DProperty = serializedObject.FindProperty(nameof(AudioManager.master2D));
-            masterVolumePersistenceProperty = serializedObject.FindProperty(nameof(AudioManager.masterVolumePersistence));
 
             channelGroupsProperty = serializedObject.FindProperty(nameof(AudioManager.channelGroups));
             defaultChannelGroupProperty = serializedObject.FindProperty(nameof(AudioManager.defaultChannelGroup));
+
+            persistenceKeyProperty = serializedObject.FindProperty(nameof(AudioManager.persistenceKey));
+            masterVolumePersistenceProperty = serializedObject.FindProperty(nameof(AudioManager.masterVolumePersistence));
 
             //channelAudioListProperty = serializedObject.FindProperty(nameof(AudioManager.channelAudio));
             //channelNameListProperty = serializedObject.FindProperty(nameof(AudioManager.channelNames));
@@ -158,7 +162,12 @@ namespace Texel
             EditorGUILayout.PropertyField(masterVolumeProperty, new GUIContent("Master Volume", "The default master volume. Can be overridden locally by users."));
             EditorGUILayout.PropertyField(masterMuteProperty, new GUIContent("Master Mute", "Whether all audio is muted by default"));
             //EditorGUILayout.PropertyField(master2DProperty, new GUIContent("Master 2D", "Whether the default spatial audio mode is 2D"));
-            EditorGUILayout.PropertyField(masterVolumePersistenceProperty, new GUIContent("Persist Master Volume", "Whether the master volume setting is saved between sessions for each user"));
+            
+
+            EditorGUILayout.Space();
+            EditorGUILayout.LabelField("Persistence", EditorStyles.boldLabel);
+            EditorGUILayout.PropertyField(persistenceKeyProperty, new GUIContent("Persistence Key", "A value unique to this video player for this world."));
+            EditorGUILayout.PropertyField(masterVolumePersistenceProperty, new GUIContent("Persist Master", "Whether the master volume and mute setting is saved between sessions for each user"));
 
             EditorGUILayout.Space();
             EditorGUILayout.LabelField("Audio Profiles", EditorStyles.boldLabel);