+++ Packages/com.texelsaur.video/Editor/Inspectors/Component/URL Source/PlaylistQueueInspector.cs
@@ -6,8 +6,15 @@ namespace Texel
     [CustomEditor(typeof(PlaylistQueue))]
     public class PlaylistQueueEditor : VideoUrlSourceInspector
     {
+        protected SerializedProperty allowAddProperty;
+        protected SerializedProperty addAccessProperty;
+        protected SerializedProperty allowAddFromProxyProperty;
+
         protected SerializedProperty priorityAccessProperty;
+        protected SerializedProperty allowPriorityProperty;
         protected SerializedProperty deleteAccessProperty;
+        protected SerializedProperty allowDeleteProperty;
+        protected SerializedProperty allowSelfDeleteProperty;
 
         protected SerializedProperty enableSyncQuestUrlsProperty;
         protected SerializedProperty syncTrackTitlesProperty;
@@ -18,8 +25,15 @@ namespace Texel
         {
             base.OnEnable();
 
+            allowAddProperty = serializedObject.FindProperty(nameof(PlaylistQueue.allowAdd));
+            addAccessProperty = serializedObject.FindProperty(nameof(PlaylistQueue.addAccess));
+            allowAddFromProxyProperty = serializedObject.FindProperty(nameof(PlaylistQueue.allowAddFromProxy));
+
             priorityAccessProperty = serializedObject.FindProperty(nameof(PlaylistQueue.priorityAccess));
+            allowPriorityProperty = serializedObject.FindProperty(nameof(PlaylistQueue.allowPriority));
             deleteAccessProperty = serializedObject.FindProperty(nameof(PlaylistQueue.deleteAccess));
+            allowDeleteProperty = serializedObject.FindProperty(nameof(PlaylistQueue.allowDelete));
+            allowSelfDeleteProperty = serializedObject.FindProperty(nameof(PlaylistQueue.allowSelfDelete));
 
             enableSyncQuestUrlsProperty = serializedObject.FindProperty(nameof(PlaylistQueue.enableSyncQuestUrls));
             syncTrackTitlesProperty = serializedObject.FindProperty(nameof(PlaylistQueue.syncTrackTitles));
@@ -35,8 +49,33 @@ namespace Texel
 
             EditorGUILayout.Space();
             EditorGUILayout.LabelField("Permissions", EditorStyles.boldLabel);
-            EditorGUILayout.PropertyField(priorityAccessProperty);
-            EditorGUILayout.PropertyField(deleteAccessProperty);
+
+            EditorGUILayout.PropertyField(allowAddProperty, new GUIContent("Allow Add", "Allow entries to be added to the queue."));
+            if (allowAddProperty.boolValue)
+            {
+                EditorGUI.indentLevel += 1;
+                EditorGUILayout.PropertyField(addAccessProperty, new GUIContent("Add Access", "Optional. ACL to control access to adding entries.  If not set, uses the video player's ACL settings."));
+                EditorGUILayout.PropertyField(allowAddFromProxyProperty, new GUIContent("Allow Add From Proxy", "Allows entries to be added through the input proxy, regardless of overall access control restrictions.  This applies to external systems like YouTube Search Prefab."));
+                EditorGUI.indentLevel -= 1;
+            }
+
+            EditorGUILayout.PropertyField(allowPriorityProperty, new GUIContent("Allow Priority", "Allows queue entries to be moved to the front."));
+            if (allowPriorityProperty.boolValue)
+            {
+                EditorGUI.indentLevel += 1;
+                EditorGUILayout.PropertyField(priorityAccessProperty, new GUIContent("Priority Access", "Optional. ACL to control access to the priority button.  If not set, uses the video player's ACL settings."));
+                EditorGUI.indentLevel -= 1;
+            }
+
+            EditorGUILayout.PropertyField(allowDeleteProperty, new GUIContent("Allow Delete", "Allow added queue entries to be deleted."));
+            if (allowDeleteProperty.boolValue)
+            {
+                EditorGUI.indentLevel += 1;
+                EditorGUILayout.PropertyField(deleteAccessProperty, new GUIContent("Delete Access", "Optional. ACL to control access to the delete button.  If not set, uses the video player's ACL settings."));
+                EditorGUILayout.PropertyField(allowSelfDeleteProperty, new GUIContent("Allow Self Delete", "Allows players to delete their own added entries, regardless of overall access control restrictions."));
+                EditorGUI.indentLevel -= 1;
+            }
+            
 
             EditorGUILayout.Space();
             EditorGUILayout.LabelField("Sync", EditorStyles.boldLabel);