+++ Packages/com.texelsaur.video/Runtime/Scripts/Component/URL Source/PlaylistQueue.cs
@@ -4,7 +4,6 @@ using UdonSharp;
 using UnityEngine;
 using VRC.SDK3.Data;
 using VRC.SDKBase;
-using VRC.Udon;
 
 [assembly: InternalsVisibleTo("com.texelsaur.video.Editor")]
 
@@ -21,10 +20,18 @@ namespace Texel
     {
         TXLVideoPlayer videoPlayer;
 
+        [SerializeField] protected internal bool allowAdd = true;
+        [SerializeField] protected internal AccessControl addAccess;
+        [SerializeField] protected internal bool allowAddFromProxy = false;
+
         [Tooltip("Optional. ACL to control access to the move-to-front button.  If not set, uses the video player's ACL settings.")]
         [SerializeField] protected internal AccessControl priorityAccess;
+        [SerializeField] protected internal bool allowPriority = false;
         [Tooltip("Optional. ACL to control access to the delete button.  If not set, uses the video player's ACL settings.")]
         [SerializeField] protected internal AccessControl deleteAccess;
+        //[Tooltip("Allows players to delete their own added entries, even if th")]
+        [SerializeField] protected internal bool allowDelete = false;
+        [SerializeField] protected internal bool allowSelfDelete = false;
 
         [SerializeField] protected internal bool enableSyncQuestUrls = true;
         [SerializeField] protected internal bool syncTrackTitles = true;
@@ -47,6 +54,8 @@ namespace Texel
         [UdonSynced]
         Vector3[] syncEntries; // (playlistIndex, catalogIndex, trackIndex)
         [UdonSynced]
+        int[] syncPlayerIds;
+        [UdonSynced]
         string[] syncTitles;
         [UdonSynced]
         string[] syncAuthors;
@@ -68,6 +77,10 @@ namespace Texel
         [UdonSynced]
         int syncReadyUrlUpdate = 0;
         int prevReadyUrlUpdate = 0;
+
+        [UdonSynced]
+        int syncTrackAddedUpdate = 0;
+        int prevTrackAddedUpdate = 0;
 
         bool usingQuestUrls = false;
         bool usingTitles = false;
@@ -99,6 +112,7 @@ namespace Texel
             syncUrls = new VRCUrl[0];
             syncQuestUrls = new VRCUrl[0];
             syncEntries = new Vector3[0];
+            syncPlayerIds = new int[0];
 
             // Info data
             syncTitles = new string[0];
@@ -235,12 +249,31 @@ namespace Texel
         {
             get
             {
+                if (!allowDelete)
+                    return false;
                 if (deleteAccess)
                     return deleteAccess._LocalHasAccess();
                 if (videoPlayer)
                     return videoPlayer._CanTakeControl();
                 return false;
             }
+        }
+
+        public bool _HasDeleteAccessFor(int index)
+        {
+            if (!allowDelete || index < 0 || index >= syncTrackCount)
+                return false;
+
+            bool addedSelf = syncPlayerIds[index] == Networking.LocalPlayer.playerId;
+            if (addedSelf && allowSelfDelete)
+                return true;
+
+            if (deleteAccess)
+                return deleteAccess._LocalHasAccess();
+            if (videoPlayer)
+                return videoPlayer._CanTakeControl();
+
+            return false;
         }
 
         public override bool _CanMoveNext()
@@ -402,6 +435,7 @@ namespace Texel
             {
                 syncUrls[i] = syncUrls[i + popCount];
                 syncEntries[i] = syncEntries[i + popCount];
+                syncPlayerIds[i] = syncPlayerIds[i + popCount];
                 if (usingQuestUrls)
                     syncQuestUrls[i] = syncQuestUrls[i + popCount];
                 if (usingTitles)
@@ -416,6 +450,7 @@ namespace Texel
             {
                 syncUrls[i] = VRCUrl.Empty;
                 syncEntries[i] = new Vector3(-1, -1, -1);
+                syncPlayerIds[i] = -1;
                 if (usingQuestUrls)
                     syncQuestUrls[i] = VRCUrl.Empty;
                 if (usingTitles)
@@ -466,6 +501,12 @@ namespace Texel
                 prevReadyUrlUpdate = syncReadyUrlUpdate;
                 _EventUrlReady();
             }
+
+            if (syncTrackAddedUpdate != prevTrackAddedUpdate)
+            {
+                prevTrackAddedUpdate = syncTrackAddedUpdate;
+                _UpdateInfoResolver();
+            }
         }
 
         void _PopulateResolver()
@@ -497,16 +538,29 @@ namespace Texel
 
         public bool _RemoveTrack(int index)
         {
-            if (index < 0 || index >= syncTrackCount)
+            if (!allowDelete || index < 0 || index >= syncTrackCount)
                 return false;
 
-            if (!_TakeControl(deleteAccess))
+            bool addedSelf = syncPlayerIds[index] == Networking.LocalPlayer.playerId;
+            if (addedSelf && allowSelfDelete)
+                _ForceTakeControl();
+            else if (!_TakeControl(deleteAccess))
                 return false;
 
             _PopTracks(index, 1);
             RequestSerialization();
 
             return true;
+        }
+
+        public bool _MoveTrackFront(int index)
+        {
+            if (!allowPriority)
+                return false;
+            if (!_TakeControl(priorityAccess))
+                return false;
+
+            return _MoveTrack(index, 0);
         }
 
         public bool _MoveTrack(int index, int destIndex)
@@ -518,11 +572,9 @@ namespace Texel
             if (index == destIndex)
                 return false;
 
-            if (!_TakeControl(priorityAccess))
-                return false;
-
             VRCUrl dstUrl = syncUrls[index];
             Vector3 dstEntry = syncEntries[index];
+            int dstPlayerId = syncPlayerIds[index];
             VRCUrl dstQuestUrl = VRCUrl.Empty;
             if (usingQuestUrls)
                 dstQuestUrl = syncQuestUrls[index];
@@ -542,6 +594,7 @@ namespace Texel
                 {
                     syncUrls[i] = syncUrls[i - 1];
                     syncEntries[i] = syncEntries[i - 1];
+                    syncPlayerIds[i] = syncPlayerIds[i - 1];
                     if (usingQuestUrls)
                         syncQuestUrls[i] = syncQuestUrls[i - 1];
                     if (usingTitles)
@@ -557,6 +610,7 @@ namespace Texel
                 {
                     syncUrls[i] = syncUrls[i + 1];
                     syncEntries[i] = syncEntries[i + 1];
+                    syncPlayerIds[i] = syncPlayerIds[i + 1];
                     if (usingQuestUrls)
                         syncQuestUrls[i] = syncQuestUrls[i + 1];
                     if (usingTitles)
@@ -570,6 +624,7 @@ namespace Texel
 
             syncUrls[destIndex] = dstUrl;
             syncEntries[destIndex] = dstEntry;
+            syncPlayerIds[destIndex] = dstPlayerId;
             if (usingQuestUrls)
                 syncQuestUrls[destIndex] = dstQuestUrl;
             if (usingTitles)
@@ -594,29 +649,39 @@ namespace Texel
 
         public override bool _AddTrack(VRCUrl url)
         {
-            return _AddTrack(url, VRCUrl.Empty, "");
+            return _AddTrack(url, VRCUrl.Empty, "", "", false);
+        }
+
+        public override bool _AddTrackFromProxy(VRCUrl url, VRCUrl questUrl, string title, string author)
+        {
+            return _AddTrack(url, questUrl, title, author, allowAddFromProxy);
         }
 
         public bool _AddTrack(VRCUrl url, VRCUrl questUrl, string title)
         {
-            return _AddTrack(url, questUrl, title, "");
+            return _AddTrack(url, questUrl, title, "", false);
         }
 
-        public bool _AddTrack(VRCUrl url, VRCUrl questUrl, string title, string author)
+        public bool _AddTrack(VRCUrl url, VRCUrl questUrl, string title, string author, bool force)
         {
+            if (!allowAdd)
+                return false;
             if (!URLUtil.WellFormedUrl(url))
                 return false;
 
-            if (!_TakeControl())
+            if (force)
+                _ForceTakeControl();
+            else if (!_TakeControl(addAccess))
                 return false;
 
             _EnsureSyncCapacity();
 
-            syncUrls[syncTrackCount] = url;
+            syncUrls[syncTrackCount] = url != null ? url : VRCUrl.Empty;
             if (usingQuestUrls)
-                syncQuestUrls[syncTrackCount] = questUrl;
+                syncQuestUrls[syncTrackCount] = questUrl != null ? questUrl : VRCUrl.Empty;
 
             syncEntries[syncTrackCount] = new Vector3(-1, -1, -1);
+            syncPlayerIds[syncTrackCount] = Networking.LocalPlayer.playerId;
 
             if (usingTitles || usingAuthors)
             {
@@ -699,6 +764,20 @@ namespace Texel
             return false;
         }
 
+        void _UpdateInfoResolver()
+        {
+            UrlInfoResolver resolver = videoPlayer.UrlInfoResolver;
+            if (!resolver)
+                return;
+
+            for (int i = 0; i < syncTrackCount; i++)
+            {
+                string title = usingTitles ? syncTitles[i] : null;
+                string author = usingAuthors ? syncAuthors[i] : null;
+                resolver._AddInfo(syncUrls[i], title, author);
+            }
+        }
+
         public bool _AddTrack(int playlistIndex, int catalogIndex, int trackIndex)
         {
             if (!_TakeControl())
@@ -711,6 +790,7 @@ namespace Texel
                 syncQuestUrls[syncTrackCount] = VRCUrl.Empty;
 
             syncEntries[syncTrackCount] = new Vector3(playlistIndex, catalogIndex, trackIndex);
+            syncPlayerIds[syncTrackCount] = Networking.LocalPlayer.playerId;
 
             if (usingTitles)
                 syncTitles[syncTrackCount] = "";
@@ -731,10 +811,13 @@ namespace Texel
 
         private void _EnsureSyncCapacity()
         {
+            _EnsureInit();
+
             if (syncTrackCount >= syncEntries.Length)
             {
                 syncEntries = (Vector3[])UtilityTxl.ArrayMinSize(syncEntries, syncTrackCount + 1, typeof(Vector3));
                 syncUrls = (VRCUrl[])UtilityTxl.ArrayMinSize(syncUrls, syncTrackCount + 1, typeof(VRCUrl));
+                syncPlayerIds = (int[])UtilityTxl.ArrayMinSize(syncPlayerIds, syncTrackCount + 1, typeof(int));
 
                 if (usingQuestUrls)
                     syncQuestUrls = (VRCUrl[])UtilityTxl.ArrayMinSize(syncQuestUrls, syncTrackCount + 1, typeof(VRCUrl));
@@ -751,6 +834,7 @@ namespace Texel
         {
             syncQueueUpdate += 1;
             syncTrackCount += 1;
+            syncTrackAddedUpdate += 1;
 
             bool isPlaying = videoPlayer && (videoPlayer.playerState == TXLVideoPlayer.VIDEO_STATE_LOADING || videoPlayer.playerState == TXLVideoPlayer.VIDEO_STATE_PLAYING);
 
@@ -772,6 +856,14 @@ namespace Texel
             if (!acl && videoPlayer && videoPlayer.SupportsOwnership && !videoPlayer._TakeControl())
                 return false;
 
+            if (!Networking.IsOwner(gameObject))
+                Networking.SetOwner(Networking.LocalPlayer, gameObject);
+
+            return true;
+        }
+
+        bool _ForceTakeControl()
+        {
             if (!Networking.IsOwner(gameObject))
                 Networking.SetOwner(Networking.LocalPlayer, gameObject);